Reorganise faucet files
