+++ bin/faucet/build.rs
@@ -6,7 +6,8 @@ fn main() {
     let static_dir = std::path::PathBuf::from_str(std::env!("CARGO_MANIFEST_DIR"))
         .unwrap()
         .join("src")
-        .join("static");
+        .join("server")
+        .join("resources");
     println!("cargo::rerun-if-changed={}", static_dir.to_str().expect("Valid utf-8"));
     // This makes the static files available as an embedded resource.
     static_files::resource_dir(static_dir).build().expect("Resources should build");
