+++ bin/faucet/src/config.rs
@@ -21,8 +21,8 @@ pub const DEFAULT_RPC_TIMEOUT_MS: u64 = 10000;
 #[derive(Clone, Debug, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct FaucetConfig {
-    /// Endpoint of the faucet in the format `<ip>:<port>`
-    pub endpoint: Url,
+    /// The port at which to serve the faucet's REST API.
+    pub port: u16,
     /// Node RPC gRPC endpoint in the format `http://<host>[:<port>]`
     pub node_url: Url,
     /// Timeout for RPC requests in milliseconds
@@ -37,7 +37,7 @@ impl Display for FaucetConfig {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         f.write_fmt(format_args!(
             "{{ endpoint: \"{}\", node_url: \"{}\", timeout_ms: \"{}\", asset_amount_options: {:?}, faucet_account_path: \"{}\" }}",
-            self.endpoint, self.node_url, self.timeout_ms, self.asset_amount_options, self.faucet_account_path.display()
+            self.port, self.node_url, self.timeout_ms, self.asset_amount_options, self.faucet_account_path.display()
         ))
     }
 }
@@ -45,8 +45,7 @@ impl Display for FaucetConfig {
 impl Default for FaucetConfig {
     fn default() -> Self {
         Self {
-            endpoint: Url::parse(format!("http://0.0.0.0:{DEFAULT_FAUCET_SERVER_PORT}").as_str())
-                .unwrap(),
+            port: DEFAULT_FAUCET_SERVER_PORT,
             node_url: Url::parse(format!("http://127.0.0.1:{DEFAULT_NODE_RPC_PORT}").as_str())
                 .unwrap(),
             timeout_ms: DEFAULT_RPC_TIMEOUT_MS,
@@ -54,22 +53,5 @@ impl Default for FaucetConfig {
             asset_amount_options: AssetOptions::new(vec![100, 500, 1_000]).unwrap(),
             faucet_account_path: DEFAULT_FAUCET_ACCOUNT_PATH.into(),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use tokio::net::TcpListener;
-
-    use super::FaucetConfig;
-
-    #[tokio::test]
-    async fn default_faucet_config() {
-        // Default does not panic
-        let config = FaucetConfig::default();
-        // Default can bind
-        let socket_addrs = config.endpoint.socket_addrs(|| None).unwrap();
-        let socket_addr = socket_addrs.into_iter().next().unwrap();
-        let _listener = TcpListener::bind(socket_addr).await.unwrap();
     }
 }
