+++ bin/faucet/src/faucet/distribute_fungible_asset.masm
