+++ bin/faucet/src/faucet/store.rs
