+++ bin/faucet/src/server/resources/index.css
