+++ bin/faucet/src/server/resources/index.html
