+++ bin/faucet/src/server/resources/index.js
